Address review: usize limit, tolerate missing runId
- `--limit` is now `usize` end to end (cli → runs → api). u16 was an
  artificial 16k ceiling; the server owns the real range and may raise it.
- `parse_debug_launch_body` returns `Option<String>` for the run id instead
  of hard-erroring when it is absent. The tenant-58.6 schema makes runId
  optional, and the generated `LaunchMvdResponse` deserializer on the
  documented-202 path already tolerates its absence; the fallback path
  (undocumented 200) now agrees, so a run-id-less success body is accepted
  as success (run_id: None) rather than reported as a failure.

+++ src/cli.rs
@@ -599,10 +599,10 @@ and VTIME into `runs logs` to see the surrounding logs."#
         #[arg(short = 'm', long = "match")]
         matches: Vec<String>,
 
-        /// Maximum number of events the server returns (1-999). Raise it to make
-        /// a search more exhaustive.
+        /// Maximum number of events the server returns. Raise it to make a
+        /// search more exhaustive. The server enforces the accepted range.
         #[arg(short = 'n', long, default_value = "50")]
-        limit: u16,
+        limit: usize,
 
         /// Substrings to match, as a positional alias for `-m` (all must match).
         /// At least one needle (via `-m` or here) is required.
